Rag_web-1/requirements.txt
streamlit==1.29.0
openai==1.3.7
python-dotenv==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.1
transformers==4.35.2
huggingface-hub==0.19.4
PyPDF2==3.0.1
tqdm==4.66.1
python-decouple==3.8
beautifulsoup4==4.12.2
langchain-community==0.0.16
